refactor: move workflow node params helper
