+++ packages/service/core/workflow/dispatch/utils/runtime.ts
@@ -1,0 +1,109 @@
+import { NodeInputKeyEnum } from '@fastgpt/global/core/workflow/constants';
+import {
+  FlowNodeInputTypeEnum,
+  FlowNodeTypeEnum
+} from '@fastgpt/global/core/workflow/node/constant';
+import type {
+  RuntimeNodeItemType,
+  WorkflowVariableStateLike
+} from '@fastgpt/global/core/workflow/runtime/type';
+import {
+  getReferenceVariableValue,
+  valueTypeFormat
+} from '@fastgpt/global/core/workflow/runtime/utils';
+import { nodeInputIsReference } from '@fastgpt/global/core/workflow/utils';
+import { replaceEditorVariable } from './replaceEditorVariable';
+
+/**
+ * 解析单个工作流节点运行参数。
+ *
+ * 这是调度热路径：每个节点执行前都会经过。这里按需构造 runtime variables，
+ * 并在调度层跳过静态输入的文本替换，避免无变量节点也复制整张变量表。
+ */
+export const getWorkflowNodeRunParams = ({
+  node,
+  runtimeNodesMap,
+  variableState
+}: {
+  node: RuntimeNodeItemType;
+  runtimeNodesMap: Map<string, RuntimeNodeItemType>;
+  variableState: WorkflowVariableStateLike;
+}) => {
+  if (node.flowNodeType === FlowNodeTypeEnum.pluginInput) {
+    // Format plugin input to object
+    return node.inputs.reduce<Record<string, any>>((acc, item) => {
+      acc[item.key] = valueTypeFormat(item.value, item.valueType);
+      return acc;
+    }, {});
+  }
+
+  // Dynamic input need to store a key.
+  const dynamicInput = node.inputs.find(
+    (item) => item.renderTypeList[0] === FlowNodeInputTypeEnum.addInputParam
+  );
+  const params: Record<string, any> = dynamicInput
+    ? {
+        [dynamicInput.key]: {}
+      }
+    : {};
+
+  let runtimeVariables: Record<string, unknown> | undefined;
+  const getRuntimeVariables = () => {
+    runtimeVariables ??= variableState.toRuntimeRecord();
+    return runtimeVariables;
+  };
+
+  node.inputs.forEach((input) => {
+    // Special input, not format
+    if (input.key === dynamicInput?.key) return;
+
+    // Skip some special key
+    if (
+      [NodeInputKeyEnum.childrenNodeIdList, NodeInputKeyEnum.httpJsonBody].includes(
+        input.key as NodeInputKeyEnum
+      )
+    ) {
+      params[input.key] = input.value;
+      return;
+    }
+
+    const rawValue = input.value;
+    const isReferenceInput = nodeInputIsReference(input);
+    const needsTextReplace = typeof rawValue === 'string' && rawValue.includes('{{');
+    let value = rawValue;
+
+    if (isReferenceInput && !needsTextReplace) {
+      value = getReferenceVariableValue({
+        value,
+        nodesMap: runtimeNodesMap,
+        variables: getRuntimeVariables(),
+        isReferenceVal: true
+      });
+    } else {
+      if (needsTextReplace) {
+        value = replaceEditorVariable({
+          text: value,
+          nodesMap: runtimeNodesMap,
+          variables: getRuntimeVariables()
+        });
+      }
+
+      if (isReferenceInput) {
+        value = getReferenceVariableValue({
+          value,
+          nodesMap: runtimeNodesMap,
+          variables: getRuntimeVariables(),
+          isReferenceVal: true
+        });
+      }
+    }
+
+    // Dynamic input is stored in the dynamic key
+    if (input.canEdit && dynamicInput && params[dynamicInput.key]) {
+      params[dynamicInput.key][input.key] = valueTypeFormat(value, input.valueType);
+    }
+    params[input.key] = valueTypeFormat(value, input.valueType);
+  });
+
+  return params;
+};
